move stripEthernetHeader into linux
