+++ pkg/networkpath/traceroute/packets/afpacket_source_linux.go
@@ -13,6 +13,9 @@ import (
 	"time"
 
 	"golang.org/x/sys/unix"
+
+	"github.com/google/gopacket"
+	"github.com/google/gopacket/layers"
 )
 
 // AFPacketSource is a PacketSource implementation using AF_PACKET.
@@ -45,16 +48,19 @@ func (a *AFPacketSource) SetReadDeadline(t time.Time) error {
 
 // Read reads a packet (starting with the IP frame)
 func (a *AFPacketSource) Read(buf []byte) (int, error) {
-	n, err := a.sock.Read(buf)
-	if err != nil {
-		return n, err
-	}
-	payload, err := stripEthernetHeader(buf[:n])
-	if err != nil {
-		return n, err
+	var payload []byte
+	for payload == nil {
+		n, err := a.sock.Read(buf)
+		if err != nil {
+			return n, err
+		}
+		payload, err = stripEthernetHeader(buf[:n])
+		if err != nil {
+			return n, err
+		}
 	}
 	copy(buf, payload)
-	return n, nil
+	return len(payload), nil
 }
 
 // Close closes the socket
@@ -65,4 +71,18 @@ func (a *AFPacketSource) Close() error {
 // htons converts a short (uint16) from host-to-network byte order.
 func htons(i uint16) uint16 {
 	return i<<8 | i>>8
+}
+
+// removes the preceding ethernet header from the buffer
+func stripEthernetHeader(buf []byte) ([]byte, error) {
+	var eth layers.Ethernet
+	err := (&eth).DecodeFromBytes(buf, gopacket.NilDecodeFeedback)
+	if err != nil {
+		return nil, fmt.Errorf("stripEthernetHeader failed to decode ethernet: %w", err)
+	}
+	// return zero bytes when the it's not an IP packet
+	if eth.EthernetType != layers.EthernetTypeIPv4 && eth.EthernetType != layers.EthernetTypeIPv6 {
+		return nil, nil
+	}
+	return eth.Payload, nil
 }
